requirements.txt
# Core dependencies (Python 3.13 compatible)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.54.0
chromadb>=0.5.0
streamlit>=1.39.0

# Document processing
python-docx==1.1.0
docx2txt==0.8
openpyxl==3.1.2
PyPDF2==3.0.1
pywin32>=307  # For .doc files on Windows (Python 3.13+)

# Text processing
tiktoken>=0.7.0  # Python 3.13 compatible

# Environment management
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1

